feat(core-platform): add a "latest" installer for each app channel
when a new version is uploaded, any installer uploaded to that version while it is the "latest"
will be uploaded to a special known path with the app name as the file name but the correct
platform specific installer extension. This shouldnt break code signing but be aware if you
are using cloudfront these "latest" files are subject to the same cache busting 15 minute
rule as other files.